Put all unshared functions in the files where they are used
That way it is not necessary to call them dead.
Reorder imports, in files where it is necessary to add or remove them
anyway.
